Fixed Android keypress code after WorldTime removal.
Tested:
Verified fplbase builds again for Android.

+++ include/fplbase/input.h
@@ -517,7 +517,7 @@ class InputSystem {
   bool relative_mouse_mode_;
 
   // How long since we've sent a keypress event to keep the CPU alive.
-  int32_t last_android_keypress_;
+  double last_android_keypress_;
 # endif
 };
